+++ modules/kubernetes/aks-core/modules.tf
@@ -177,19 +177,19 @@ module "aad_pod_identity" {
 }
 
 # AZ Metrics
-module "az_metrics" {
+module "azure_metrics" {
   depends_on = [module.opa_gatekeeper, module.aad_pod_identity]
 
   for_each = {
-    for s in ["az-metrics"] :
+    for s in ["azure-metrics"] :
     s => s
-    if var.az_metrics_enabled
+    if var.azure_metrics_enabled
   }
 
-  source = "../../kubernetes/az-metrics"
+  source = "../../kubernetes/azure-metrics"
 
-  client_id       = var.az_metrics_config.client_id
-  resource_id     = var.az_metrics_config.resource_id
+  client_id       = var.azure_metrics_config.client_id
+  resource_id     = var.azure_metrics_config.resource_id
   subscription_id = data.azurerm_client_config.current.subscription_id
 }