+++ modules/kubernetes/aks-core/variables.tf
@@ -386,13 +386,13 @@ variable "kube_state_metrics_namespaces_extras" {
   default     = []
 }
 
-variable "az_metrics_enabled" {
+variable "azure_metrics_enabled" {
   description = "Should AZ Metrics be enabled"
   type        = bool
   default     = true
 }
 
-variable "az_metrics_config" {
+variable "azure_metrics_config" {
   description = "AZ Metrics configuration"
   type = object({
     client_id   = string